chore : model 패키지 하위에 user, transaction과 token패키지 생성 및 이동
