+++ server/src/main/java/com/watchtogether/server/users/domain/model/transaction/TransactionWithdraw.java
@@ -1,44 +1,31 @@
-package com.watchtogether.server.users.domain.model;
+package com.watchtogether.server.users.domain.model.transaction;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.time.LocalDateTime;
-import javax.validation.constraints.Email;
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-public class TransactionCharge {
+public class TransactionWithdraw {
 
     @Getter
     @NoArgsConstructor
     @AllArgsConstructor
-    @Schema(name = "TransactionChargeRequest", description = "캐쉬 충전 요청")
+    @Schema(name = "TransactionWithdrawRequest", description = "파티 참가시 선결제 요청")
     public static class Request {
 
-        @Email
-        @NotBlank(message = "이메일을 입력해 주세요.")
-        @Schema(description = "아이디", example = "abc@ikdmd.kg.lr")
-        private String email;
-
-        @NotBlank(message = "비밀번호를 입력해 주세요.")
-        @Schema(description = "비밀번호")
-        private String password;
-
-        @NotNull(message = "생년월일을 입력해 주세요.")
-        @Schema(description = "충전 금액")
-        private Long amount;
+        @Schema(description = "party 아이디")
+        private Long partyId;
 
     }
 
     @Getter
     @NoArgsConstructor
     @AllArgsConstructor
-    @Schema(name = "SignUpResponse", description = "사용자 회원가입 응답")
+    @Schema(name = "TransactionWithdrawResponse", description = "파티 참가시 선결제 응답")
     public static class Response {
 
-        @Schema(description = "닉네임")
+        @Schema(description = "닉네임", example = "abc@ikdmd.kg.lr")
         private String nickname;
 
         @Schema(description = "거래 종류")
@@ -53,7 +40,7 @@ public class TransactionCharge {
         @Schema(description = "거래 후 잔액")
         private Long balanceSnapshot;
 
-        @Schema(description = "거래자 닉네임")
+        @Schema(description = "거래자 이메일")
         private String traderNickname;
 
         @Schema(description = "거래 일시")
