+++ server/src/main/java/com/watchtogether/server/users/domain/model/user/ChangePassword.java
@@ -1,4 +1,4 @@
-package com.watchtogether.server.users.domain.model;
+package com.watchtogether.server.users.domain.model.user;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import javax.validation.constraints.NotBlank;
@@ -7,15 +7,15 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-public class CheckPassword {
+public class ChangePassword {
 
     @Getter
     @NoArgsConstructor
     @AllArgsConstructor
-    @Schema(name = "CheckPasswordRequest", description = "사용자 현재 비밀번호 검사 요청")
+    @Schema(name = "ChangePasswordRequest", description = "새로운 비밀번호로 변경 요청")
     public static class Request {
 
-        @NotBlank(message = "비밀번호를 입력해 주세요.")
+        @NotBlank(message = "새로운 비밀번호를 입력해 주세요.")
         @Schema(description = "비밀번호")
         private String password;
 
@@ -25,7 +25,7 @@ public class CheckPassword {
     @Getter
     @NoArgsConstructor
     @AllArgsConstructor
-    @Schema(name = "CheckPasswordResponse", description = "사용자 현재 비밀번호 검사 응답")
+    @Schema(name = "CheckPasswordResponse", description = "새로운 비밀번호로 변경 응답")
     public static class Response {
 
         @Schema(description = "성공 응답 메시지")
