+++ server/src/main/java/com/watchtogether/server/users/domain/model/user/SearchUser.java
@@ -1,19 +1,24 @@
-package com.watchtogether.server.users.domain.model;
+package com.watchtogether.server.users.domain.model.user;
+
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-public class SignOutUser {
+public class SearchUser {
 
     @Getter
     @NoArgsConstructor
     @AllArgsConstructor
-    @Schema(name = "SignOutResponse", description = "사용자 로그아웃 응답")
     public static class Response {
 
-        @Schema(description = "메시지")
+        @Schema(description = "닉네임")
+        private String nickname;
+
+        @Schema(description = "성공 메시지")
         private String message;
+
     }
+
 }
