+++ server/src/main/java/com/watchtogether/server/users/domain/model/user/VerifyEmail.java
@@ -1,4 +1,4 @@
-package com.watchtogether.server.users.domain.model;
+package com.watchtogether.server.users.domain.model.user;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
