+++ server/src/test/java/com/watchtogether/server/users/service/impl/UserServiceImplTest.java
@@ -9,8 +9,8 @@ import static org.mockito.Mockito.verify;
 import com.watchtogether.server.components.mail.MailComponents;
 import com.watchtogether.server.users.domain.dto.UserDto;
 import com.watchtogether.server.users.domain.entitiy.User;
-import com.watchtogether.server.users.domain.model.SignInUser;
-import com.watchtogether.server.users.domain.model.SignUpUser;
+import com.watchtogether.server.users.domain.model.user.SignInUser;
+import com.watchtogether.server.users.domain.model.user.SignUpUser;
 import com.watchtogether.server.users.domain.repository.UserRepository;
 import com.watchtogether.server.users.domain.type.UserStatus;
 import java.time.LocalDate;
